Create Validation-JSObjectSubmission
Changing the message to the user to a floating message box and adding an object to submit. This could add to the service happy provides... maybe changing the scope a bit, but since Happy provided the mechanism, I took it s step further to use it to submit the form in its entirety.